converter/ICDAR2VOC/requirements.txt
numpy==1.14.3+mkl
opencv_python==3.4.0.12
Pillow==5.1.0
